CommunicationServer renamed to SerialPortPeripheral and other files have been changed accordingly. Also modified attempt connection to go through TelemetryReporting instead of directly to the Peripherals layer.
