--- src/BusinessLayer/CommunicationServer.h
@@ -1,16 +1,0 @@
-#pragma once
-class QString;
-class QSerialPort;
-
-/*
- *
- */
-class CommunicationServer
-{
-public:
-    CommunicationServer(QSerialPort& serialPort, QString portName);
-    bool attemptConnection();
-private:
-    QSerialPort& serialPort_;
-
-};
