+++ src/Peripherals/SerialPortPeripheral.h
@@ -1,11 +1,15 @@
-#ifndef SERIALPORTPERIPHERAL_H
-#define SERIALPORTPERIPHERAL_H
+#pragma once
+class QString;
+class QSerialPort;
 
-
+/*
+ *
+ */
 class SerialPortPeripheral
 {
 public:
-    SerialPortPeripheral();
+    SerialPortPeripheral(QSerialPort& serialPort, QString portName);
+    bool attemptConnection();
+private:
+    QSerialPort& serialPort_;
 };
-
-#endif // SERIALPORTPERIPHERAL_H
