+++ src/telemTestTool.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 3.5.1, 2016-03-06T09:53:43. -->
+<!-- Written by QtCreator 3.5.1, 2016-03-21T12:23:16. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
